Panaversity Class Learning/10_Tool Calling Patterns/pyproject.toml
[project]
name = "10-tool-calling-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.45.0",
    "openai-agents>=0.4.0",
]
